+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 - Added `Env.read_env` to read into the environment from a .env file.
 - Added the `optional` env argument that defaults to acting as a noop.
+- Made the default settings accessible directly from the class.
 
 ## [0.1.1] - Unreleased